[Feat] #19 - UIImage Extension 코드 분리, 새로운 파일로 이동
