+++ Carmunication/Controllers/SessionListViewController.swift
@@ -217,26 +217,6 @@ extension SessionListViewController {
     }
 }
 
-extension UIImage {
-
-    public static func pixel(ofColor color: UIColor) -> UIImage {
-        let pixel = CGRect(x: 0.0, y: 0.0, width: 1.0, height: 1.0)
-        UIGraphicsBeginImageContext(pixel.size)
-
-        defer {
-            UIGraphicsEndImageContext()
-        }
-
-        guard let context = UIGraphicsGetCurrentContext() else {
-            return UIImage()
-        }
-        context.setFillColor(color.cgColor)
-        context.fill(pixel)
-
-        return UIGraphicsGetImageFromCurrentImageContext() ?? UIImage()
-    }
-}
-
 // MARK: - Preview canvas 세팅
 import SwiftUI
