+++ Carmunication/Extensions/UIImage+Extension.swift
@@ -1,0 +1,27 @@
+//
+//  UIImage+Extension.swift
+//  Carmunication
+//
+//  Created by 김동현 on 2023/10/03.
+//
+import UIKit
+
+extension UIImage {
+
+    public static func pixel(ofColor color: UIColor) -> UIImage {
+        let pixel = CGRect(x: 0.0, y: 0.0, width: 1.0, height: 1.0)
+        UIGraphicsBeginImageContext(pixel.size)
+
+        defer {
+            UIGraphicsEndImageContext()
+        }
+
+        guard let context = UIGraphicsGetCurrentContext() else {
+            return UIImage()
+        }
+        context.setFillColor(color.cgColor)
+        context.fill(pixel)
+
+        return UIGraphicsGetImageFromCurrentImageContext() ?? UIImage()
+    }
+}
